// src/codegen/DaoPreProposeSingle.types.ts
export type Uint128 = string;
export type DepositToken = {
  token: {
    denom: UncheckedDenom;
  };
} | {
  voting_module_token: {
    token_type: VotingModuleTokenType;
  };
};
export type UncheckedDenom = {
  native: string;
} | {
  cw20: string;
};
export type VotingModuleTokenType = "native" | "cw20";
export type DepositRefundPolicy = "always" | "only_passed" | "never";
export type PreProposeSubmissionPolicy = {
  anyone: {
    denylist: Addr[];
  };
} | {
  specific: {
    allowlist: Addr[];
    dao_members: boolean;
    denylist: Addr[];
  };
};
export type Addr = string;
export interface InstantiateMsg {
  deposit_info?: UncheckedDepositInfo | null;
  extension: Empty;
  submission_policy: PreProposeSubmissionPolicy;
}
export interface UncheckedDepositInfo {
  amount: Uint128;
  denom: DepositToken;
  refund_policy: DepositRefundPolicy;
}
export interface Empty {
  [k: string]: unknown;
}